+++ src/components/choicefield/choicefieldTypeEnum.ts
@@ -1,0 +1,17 @@
+'use strict';
+
+/**
+ * @ngdoc enum
+ * @name choicefieldType
+ * @module officeuifabric.components.choicefield
+ * 
+ * @readonly
+ * 
+ * @description
+ * This attribute can be set on the uif-choicefield-option directive and determines whether it
+ * is a radio button or a checkbox
+ */
+export enum ChoicefieldType {
+    radio,
+    checkbox
+};