+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 #### Features
 
+- **uif-choicefield**: directive added (by @rolandoldengarm; closes #16)
 - **uif-icon**: added validation to specified icon (by @andrewconnell)
 - **uif-link**: directive added (by @tdwhite0; closes #23)
 - **uif-overlay**: directive added (by @johnmeilleur; closes #28)